requirements.txt
-i
aiohttp==1.0.5;
async-timeout==3.0.0;
certifi==2018.4.16
chardet==3.0.4
click==6.7
discord.py==0.16.12
idna==2.7
mongoengine==0.15.3
multidict==4.3.1;
munch==2.3.2
pymongo==3.7.1
pyyaml==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23;
websockets==3.4
